Fix OrganizationServiceServlet for non-root users
In VIVO 1.7, the vitro:hiddenFromDisplayBelowRoleLevelAnot for
obo:BFO_0000051 (has part) is 'nobody'. This means that only the root
user can see it. This broke the OrganizationServiceServlet, which
outputs json of the org tree by using that predicate.

The quick and dirty solution is to allow all users to display that
predicate.